requirements.txt
line-bot-sdk==3.*
Flask
mysql-connector-python
google-cloud-secret-manager
google.generativeai